requirements.txt
asgiref==3.4.1
boto3==1.21.0
botocore==1.24.0
dataclasses==0.8
Django==3.2.11
django-cors-headers==3.10.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
djongo==1.3.6
dnspython==2.2.0
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.4.0
PyJWT==2.3.0
pymongo==3.12.1
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.1
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.0.1
urllib3==1.26.8
whitenoise==5.3.0